code cleaned

+++ source/likelihoodsxx/system_utils.cpp
@@ -26,7 +26,7 @@ void decouple(
 void printvec_s(const char * name, const vec_s & v)
 {
     printf("\n");
-    for(int i = 0; i < v.size(); ++i)
+    for(size_t i = 0; i < v.size(); ++i)
         printf("%s[%d]: %lf\n", name, i, v[i].val());
     printf("\n");
 }
@@ -34,7 +34,7 @@ void printvec_s(const char * name, const vec_s & v)
 void printvec_d(const char * name, const vec_d & v)
 {
     printf("\n");
-    for(int i = 0; i < v.size(); ++i)
+    for(size_t i = 0; i < v.size(); ++i)
         printf("%s[%d]: %lf\n", name, i, v[i]);
     printf("\n");
 }
